requirements.txt
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
matplotlib==3.9.2
seaborn==0.13.2
missingno==0.5.2
plotly==5.23.0
statsmodels==0.14.2
umap-learn==0.5.6
numba==0.59.1
llvmlite==0.42.0
kaleido
